Refactor of createClass
